[add]　MVC設計の導入
