+++ flaskdb/controller/view/static/js/flaskdb.js
