+++ flaskdb/controller/view/templates/additem.html
@@ -3,7 +3,7 @@
 {% block head %}
   <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>flaskdb | Search Item</title>
+  <title>flaskdb | Add Item</title>
   <link href="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet">
   <link href="/static/css/flaskdb.css" rel="stylesheet">
   <!-- A Sample Web-DB Application for DB-DESIGN lecture -->
@@ -25,11 +25,11 @@
     <div class="navbar-collapse collapse">
       <ul class="nav navbar-nav">
         {% if not "username" in session %}
-        <li><a href="{{ url_for('app.login') }}">Login</a></li>
+        <li><a href="{{ url_for('auth.login') }}">Login</a></li>
         {% else %}
-        <li><a href="{{ url_for('app.additem') }}">Add Item</a></li>
-        <li><a href="{{ url_for('app.searchitem') }}">Search Item</a></li>
-        <li><a href="{{ url_for('app.logout') }}">Logout</a></li>
+        <li><a href="{{ url_for('item.additem') }}">Add Item</a></li>
+        <li><a href="{{ url_for('item.searchitem') }}">Search Item</a></li>
+        <li><a href="{{ url_for('auth.logout') }}">Logout</a></li>
         {% endif %}
       </ul>
     </div>
@@ -39,7 +39,7 @@
 
 {% block content %}
 <div class="container">
-  <div class="page-header"><h1>Search Item</h1></div>
+  <div class="page-header"><h1>Add Item</h1></div>
   {% with messages = get_flashed_messages(with_categories=true) %}
     {% if messages %}
       {% for category, message in messages %}
@@ -54,10 +54,14 @@
       {% endfor %}
     {% endfor %}
   {% endif %}
-  <form action="{{ url_for('app.searchitem') }}" method="POST" novalidate class="mb-10">
+  <form action="{{ url_for('item.additem') }}" method="POST" novalidate class="mb-10">
     <div class="form-group">
       <label for="itemname">{{ form.itemname.label }}</label>
-      {{ form.itemname(type="text", id="itemname", class="form-control", placeholder="Enter keyword") }}
+      {{ form.itemname(type="text", id="itemname", class="form-control", placeholder="Enter item name") }}
+    </div>
+    <div class="form-group">
+      <label for="price">{{ form.price.label }}</label>
+      {{ form.price(type="text", id="price", class="form-control", placeholder="Enter price") }}
     </div>
     {{ form.csrf_token }}
     {{ form.cancel(class="btn") }}
