+++ flaskdb/controller/view/templates/receive.html
